add Pry::Testable, and friends.
This commit replaces lib/pry/test/helper.rb by using category modules.
There's now:

* Pry::Testable:
  * Pry::Testable::Evalable
  * Pry::Testable::Mockable
  * Pry::Testable::Utility
  * Pry::Testable::PryTester

'include Pry::Testable' includes all of the above.
For the pry test suite it worked out best to include 'Pry::Testable' at the top-level.

In plugins I've written though this hasn't been the case, and I only normally need:
Pry::Testable::Evalable. So it reduces pollution for the third-party cases, which this
code is intended for since it lives within lib/ of the project.

What do you think?
