+++ spec/helper.rb
@@ -1,5 +1,5 @@
 require 'bundler/setup'
-require 'pry/test/helper'
+require 'pry/testable'
 Bundler.require :default, :test
 require_relative 'spec_helpers/mock_pry'
 require_relative 'spec_helpers/repl_tester'
@@ -26,12 +26,12 @@ if ENV["SET_TRACE_FUNC"]
   }
 end
 
-puts "Ruby v#{RUBY_VERSION} (#{defined?(RUBY_ENGINE) ? RUBY_ENGINE : "ruby"}), Pry v#{Pry::VERSION}, method_source v#{MethodSource::VERSION}, CodeRay v#{CodeRay::VERSION}, Pry::Slop v#{Pry::Slop::VERSION}"
-
 RSpec.configure do |config|
   config.expect_with :rspec do |c|
     c.syntax = [:should, :expect]
   end
-
-  config.include PryTestHelpers
+  config.before(:each) { Pry::Testable.prepare_config! }
+  include Pry::Testable
 end
+
+puts "Ruby v#{RUBY_VERSION} (#{defined?(RUBY_ENGINE) ? RUBY_ENGINE : "ruby"}), Pry v#{Pry::VERSION}, method_source v#{MethodSource::VERSION}, CodeRay v#{CodeRay::VERSION}, Pry::Slop v#{Pry::Slop::VERSION}"
